requirements.txt
Babel==2.6.0
Django==2.1.2
django-mathfilters==0.4.0
mysqlclient==1.3.13
pytz==2018.5
six==1.11.0
djangorestframework
pygments
